Remove bundler usage
