+++ packages/jasmine-parameterized/package.json
@@ -1,28 +1,35 @@
 {
     "name": "@ni/jasmine-parameterized",
-    "version": "1.0.0",
+    "version": "0.1.0",
     "description": "A utility to write parameterized jasmine tests",
     "scripts": {
-        "build": "npm run compile",
-        "compile": "tsc -b ./tsconfig.cjs.json ./tsconfig.esm.json ./tsconfig.types.json",
+        "build": "tsc -b",
         "lint": "eslint .",
         "format": "eslint . --fix",
         "pack": "npm pack",
         "invoke-publish": "npm publish",
         "tdd": "npm run build && npm run test",
-        "test": "npm run test:commonjs && npm run test:esm",
-        "test:commonjs": "jasmine",
-        "test:esm": "karma start karma.conf.js --browsers=ChromeHeadlessOpt --single-run"
+        "test": "npm run test:node && npm run test:browser",
+        "test:node": "jasmine",
+        "test:browser": "karma start karma.conf.cjs --browsers=ChromeHeadlessOpt --single-run",
+        "test:browser:debugger": "karma start karma.conf.cjs --browsers=ChromeDebugging"
     },
     "repository": {
         "type": "git",
         "url": "git+https://github.com/ni/nimble.git",
         "directory": "packages/jasmine-parameterized"
     },
-    "types": "./dist/types/index.d.ts",
+    "publishConfig": {
+        "access": "public"
+    },
+    "type": "module",
     "exports": {
-        "import": "./dist/esm/index.js",
-        "require": "./dist/commonjs/index.js"
+        ".": {
+            "import": {
+                "types": "./dist/esm/index.d.ts",
+                "default": "./dist/esm/index.js"
+            }
+        }
     },
     "author": "National Instruments",
     "license": "MIT",
@@ -30,18 +37,13 @@
         "@ni/eslint-config-javascript": "^4.2.0",
         "@ni/eslint-config-typescript": "^4.2.0",
         "@types/jasmine": "^4.3.1",
-        "@types/yargs": "^17.0.10",
         "jasmine": "^4.2.0",
         "jasmine-core": "^4.5.0",
         "karma": "^6.3.0",
         "karma-jasmine": "^5.1.0",
         "karma-jasmine-html-reporter": "^2.0.0",
-        "karma-source-map-support": "^1.4.0",
-        "karma-sourcemap-loader": "^0.3.7",
         "karma-spec-reporter": "^0.0.36",
-        "karma-webpack": "^5.0.0",
         "playwright": "^1.30.0",
-        "source-map-loader": "^4.0.1",
         "typescript": "~4.8.2"
     }
 }
