+++ packages/jasmine-parameterized/spec/support/jasmine.json
@@ -1,10 +1,7 @@
 {
     "spec_dir": ".",
     "spec_files": [
-      "dist/commonjs/**/*[sS]pec.?(m)js"
-    ],
-    "helpers": [
-      "helpers/**/*.?(m)js"
+      "dist/esm/**/*.spec.js"
     ],
     "env": {
       "stopSpecOnExpectationFailure": false,
