--- packages/jasmine-parameterized/tsconfig.cjs.json
@@ -1,7 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-      "outDir": "./dist/commonjs",
-      "module": "commonjs"
-    }
-}
