--- packages/jasmine-parameterized/tsconfig.esm.json
@@ -1,7 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "outDir": "./dist/esm",
-        "module": "ES2020",
-    }
-}
