--- packages/jasmine-parameterized/tsconfig.types.json
@@ -1,8 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "outDir": "./dist/types",
-        "declaration": true,
-        "emitDeclarationOnly": true
-    }
-}
